added list transfer functionlity & updated component structure
