+++ app/components/ProgressBar.tsx
@@ -13,7 +13,7 @@ export default function ProgressBar({ listID }: ProgressBarProps) {
 
   return (
     <>
-      <div className=" bg-grey w-[95%] m-auto h-[8px] mb-3 block rounded-md">
+      <div className=" bg-grey w-[95%] m-auto h-[8px] mb-3 block rounded-md ">
         <div
           className={` bg-brand h-[8px]  mb-2 block rounded-md transition-all delay-50 duration-500 ease-in-out`}
           style={{ width: `${progress}%` }}
