+++ app/listTransfer/index.ts
@@ -1,6 +1,9 @@
 import JSZip from "jszip";
 import { databaseAPI } from "../database/api/api";
 import { saveAs } from "file-saver";
+import { Dispatch, SetStateAction } from "react";
+import Ajv, { JSONSchemaType, DefinedError } from "ajv";
+import { schema } from "./schema";
 
 const handleListExport = async (listId: string) => {
   // create the file object
@@ -23,12 +26,14 @@ const handleListExport = async (listId: string) => {
       // see FileSaver.js
       saveAs(content, `${list.name}.zip`);
     });
-    // zip file
-    // allow the user to download the fil}
   }
 };
 
-const handleFileImport = async (e: React.ChangeEvent<HTMLInputElement>) => {
+const handleFileImport = async (
+  e: React.ChangeEvent<HTMLInputElement>,
+  setFileName: Dispatch<SetStateAction<string>>,
+  setIsImport: Dispatch<SetStateAction<boolean>>
+) => {
   if (e.target.files) {
     const file = e.target.files[0];
     if (file) {
@@ -39,19 +44,38 @@ const handleFileImport = async (e: React.ChangeEvent<HTMLInputElement>) => {
           const zip = await JSZip.loadAsync(zipData ?? "");
           zip.forEach(async (relPath, ZipEntry) => {
             console.log("file: ", relPath);
-            const content = await ZipEntry.async("text");
+            setFileName(relPath);
+            console.log(relPath);
+            const fileData = await ZipEntry.async("text");
+            console.log(JSON.parse(fileData));
+            const importedList = validateJSON(fileData);
+
             // TODO: add to data to database
+            if (importedList !== undefined)
+              databaseAPI.addImportedList(importedList);
             // TODO: let users know it has been added,or deleted
             // TODO:  redirect to homepage
-            console.log(JSON.parse(content));
           });
         } catch (err) {
           console.error("Error loading ZIP file:", err);
         }
       };
       reader.readAsArrayBuffer(file);
+      setIsImport(false);
     }
   }
 };
 
+const ajv = new Ajv();
+const validate = ajv.compile(schema);
+
+function validateJSON(data: string) {
+  const list = JSON.parse(data);
+  if (validate(list)) {
+    console.log("validated data", list);
+    return list;
+  } else {
+    console.log(validate.errors);
+  }
+}
 export { handleListExport, handleFileImport };
