+++ package.json
@@ -3,15 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build --turbopack",
+    "dev": "next dev",
+    "build": "next build",
     "start": "next start",
-    "lint": "eslint"
+    "lint": "eslint",
+    "lint:fix": "eslint --fix"
   },
   "dependencies": {
     "@headlessui/react": "^2.2.7",
     "@tailwindcss/postcss": "^4.1.12",
     "@types/file-saver": "^2.0.7",
+    "ajv": "^7.0.4",
     "dexie": "^4.2.0",
     "dexie-react-hooks": "^4.2.0",
     "file-saver": "^2.0.5",
@@ -21,17 +23,20 @@
     "postcss": "^8.5.6",
     "react": "19.1.0",
     "react-dom": "19.1.0",
+    "react-hot-toast": "^2.6.0",
     "react-tiny-popover": "^8.1.6",
     "tailwindcss": "^4.1.12",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@opennextjs/cloudflare": "^1.8.3",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
     "eslint-config-next": "15.5.2",
-    "typescript": "^5"
+    "typescript": "^5",
+    "wrangler": "^4.38.0"
   }
 }
